Splits common config in a Base class used from both unit and integration tests.
